Rename module containing api client
